requirements.txt
python-youtube
python-dotenv
isodate
discord-webhook